setting up basic templates for survey - google sheets figured out
each "view" will consist of one or many questions &/or textareas to fill be filled out by the user.
the backend is a google sheet.
the service to send to google sheets is sheet.best -> see index.js for more info on how that works.

a survey object is created to store the values from each of the questions
some answers to questions will produce different view - this functionality for navigating between views will come later.

index.js is still full of functions from an earlier version of this project.